Authorization and bug fixes

+++ WebShopREST/src/services/DiskService.java
@@ -1,6 +1,7 @@
 package services;
 
 import java.util.ArrayList;
+
 import java.util.HashMap;
 
 import javax.servlet.http.HttpServletRequest;
@@ -15,10 +16,9 @@ import javax.ws.rs.core.Response;
 
 import helpers.HelperMethods;
 import model.Disk;
+import model.Role;
 import model.User;
 import repository.DiskRepository;
-import repository.UserRepository;
-import repository.VmRepository;
 
 
 @Path("/disks")
@@ -27,11 +27,19 @@ public class DiskService {
 	@Context
 	HttpServletRequest request;
 
+	User loggedUser;
+	
 	@POST
 	@Path("/add")
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response addDisk(HashMap<String, String> data) {
+		
+		loggedUser = (User) request.getSession().getAttribute("user");
+		
+		if (loggedUser == null || loggedUser.getRole().equals(Role.user)) {
+			return Response.status(403).entity(HelperMethods.GetJsonValue("Unauthorized")).build();
+		}
 		if (!DiskRepository.isUniqueDisk(data.get("name"))) {
 			return Response.status(200).entity(HelperMethods.GetJsonValue("existError")).build();
 		}
@@ -39,7 +47,7 @@ public class DiskService {
 		if (!success) {
 			return Response.status(400).entity("Error adding disk").build();
 		}
-		return Response.status(200).entity(HelperMethods.GetJsonValue(success)).build();
+		return Response.status(200).entity(HelperMethods.GetJsonValue("success")).build();
 
 	}
 
@@ -47,6 +55,11 @@ public class DiskService {
 	@Path("/getAll")
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response getDisks() {
+		loggedUser = (User) request.getSession().getAttribute("user");
+		
+		if (loggedUser == null || !loggedUser.getRole().equals(Role.superadmin)) {
+			return Response.status(403).entity(HelperMethods.GetJsonValue("Unauthorized")).build();
+		}
 		ArrayList<Disk> disks = DiskRepository.getDisks();
 
 		return Response.status(200).entity(HelperMethods.GetJsonValue(disks)).build();
@@ -58,6 +71,13 @@ public class DiskService {
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response delete(HashMap<String, String> data) {
+		
+		loggedUser = (User) request.getSession().getAttribute("user");
+		
+		if (loggedUser == null || loggedUser.getRole().equals(Role.user)) {
+			return Response.status(403).entity(HelperMethods.GetJsonValue("Unauthorized")).build();
+		}
+		
 		String name = data.get("name");
 		if (DiskRepository.deleteDisk(name)) {
 			return Response.status(200).entity(DiskRepository.getDisks()).build();
@@ -71,6 +91,11 @@ public class DiskService {
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response searchDisks(HashMap<String, String> data) {
+		loggedUser = (User) request.getSession().getAttribute("user");
+		
+		if (loggedUser == null || !loggedUser.getRole().equals(Role.superadmin)) {
+			return Response.status(403).entity(HelperMethods.GetJsonValue("Unauthorized")).build();
+		}
 		String name = data.get("name");
 		return Response.status(200).entity(DiskRepository.searchDisks(name)).build();
 
@@ -81,6 +106,11 @@ public class DiskService {
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response getbyName(HashMap<String, String> data) {
+		loggedUser = (User) request.getSession().getAttribute("user");
+		
+		if (loggedUser == null) {
+			return Response.status(403).entity(HelperMethods.GetJsonValue("Unauthorized")).build();
+		}
 		Disk disk = DiskRepository.findByName(data.get("name"));
 		if (disk == null)
 			return Response.status(400).entity(HelperMethods.GetJsonValue("Error getting disk")).build();
@@ -93,8 +123,13 @@ public class DiskService {
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response updateDisk(HashMap<String, String> data) {
+		loggedUser = (User) request.getSession().getAttribute("user");
+		
+		if (loggedUser == null || loggedUser.getRole().equals(Role.user)) {
+			return Response.status(403).entity(HelperMethods.GetJsonValue("Unauthorized")).build();
+		}
 		if (!data.get("oldName").equals(data.get("name")) && !DiskRepository.isUniqueDisk(data.get("name"))) {
-			return Response.status(200).entity("existError").build();
+			return Response.status(200).entity(HelperMethods.GetJsonValue("existError")).build();
 		}
 		boolean success = DiskRepository.updateDisk(data);
 		if (!success) {
@@ -109,6 +144,11 @@ public class DiskService {
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response filterDisks(HashMap<String, String> data) {
+		loggedUser = (User) request.getSession().getAttribute("user");
+		
+		if (loggedUser == null || !loggedUser.getRole().equals(Role.superadmin)) {
+			return Response.status(403).entity(HelperMethods.GetJsonValue("Unauthorized")).build();
+		}
 		return Response.status(200).entity(DiskRepository.filterDisks(data)).build();
 	}
 	
@@ -117,10 +157,14 @@ public class DiskService {
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response getbyOrganization(HashMap<String, String> data) {
 		
-		User user = (User) request.getSession().getAttribute("user");
+		loggedUser = (User) request.getSession().getAttribute("user");
+		
+		if (loggedUser == null || loggedUser.getRole().equals(Role.superadmin)) {
+			return Response.status(403).entity(HelperMethods.GetJsonValue("Unauthorized")).build();
+		}
 
 
-		return Response.status(200).entity(DiskRepository.getDisksByCompany(user.getOrganization())).build();
+		return Response.status(200).entity(DiskRepository.getDisksByCompany(loggedUser.getOrganization())).build();
 
 	}
