+++ WebShopREST/src/services/UserService.java
@@ -1,6 +1,7 @@
 package services;
 
 import java.util.HashMap;
+
 import javax.servlet.http.HttpServletRequest;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.GET;
@@ -11,18 +12,18 @@ import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import helpers.HelperMethods;
+import model.Role;
 import model.User;
-import model.VMcategory;
-import repository.DiskRepository;
 import repository.UserRepository;
-import repository.VmCategoryRepository;
-import repository.VmRepository;
+
 
 @Path("/users")
 public class UserService {
 
 	@Context
 	HttpServletRequest request;
+	
+	User loggedUser;
 
 	public UserService() {
 
@@ -35,7 +36,7 @@ public class UserService {
 	public Response login(HashMap<String, String> data) {
 		User user = UserRepository.login(data.get("email"), data.get("password"));
 		if (user == null) {
-			return Response.status(400).entity("Invalid username and/or password").build();
+			return Response.status(400).entity(HelperMethods.GetJsonValue("Invalid username and/or password")).build();
 		}
 		request.getSession().setAttribute("user", user);
 
@@ -48,9 +49,9 @@ public class UserService {
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response loggedUser() {
 
-		User user = (User) request.getSession().getAttribute("user");
-
-		return Response.status(200).entity(HelperMethods.GetJsonValue(user)).build();
+		loggedUser = (User) request.getSession().getAttribute("user");
+		
+		return Response.status(200).entity(HelperMethods.GetJsonValue(loggedUser)).build();
 
 	}
 
@@ -59,9 +60,16 @@ public class UserService {
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response register(HashMap<String, String> data) {
+		
+		loggedUser = (User) request.getSession().getAttribute("user");
+		
+		if (loggedUser == null || loggedUser.getRole().equals(Role.user)) {
+			return Response.status(403).entity(HelperMethods.GetJsonValue("Unauthorized")).build();
+		}
+		
 		String email = data.get("email");
 		if (!UserRepository.isUniqueEmail(email)) {
-			return Response.status(400).entity("Email already exists").build();
+			return Response.status(200).entity(HelperMethods.GetJsonValue("existError")).build();
 		}
 		String password = data.get("password");
 		String firstName = data.get("firstName");
@@ -89,6 +97,11 @@ public class UserService {
 	@Path("/getAll")
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response findAll() {
+		loggedUser = (User) request.getSession().getAttribute("user");
+		
+		if (loggedUser == null || !loggedUser.getRole().equals(Role.superadmin)) {
+			return Response.status(403).entity(HelperMethods.GetJsonValue("Unauthorized")).build();
+		}
 		return Response.status(200).entity(UserRepository.getUsers()).build();
 	}
 
@@ -97,6 +110,11 @@ public class UserService {
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response delete(HashMap<String, String> data) {
+		loggedUser = (User) request.getSession().getAttribute("user");
+		
+		if (loggedUser == null || loggedUser.getRole().equals(Role.user)) {
+			return Response.status(403).entity(HelperMethods.GetJsonValue("Unauthorized")).build();
+		}
 		String email = data.get("email");
 		if (UserRepository.deleteUser(email)) {
 			return Response.status(200).entity(UserRepository.getUsers()).build();
@@ -111,6 +129,11 @@ public class UserService {
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response updateMyProfile(HashMap<String, String> data) {
+		loggedUser = (User) request.getSession().getAttribute("user");
+		
+		if (loggedUser == null) {
+			return Response.status(403).entity(HelperMethods.GetJsonValue("Unauthorized")).build();
+		}
 		User loggedUser = (User) request.getSession().getAttribute("user");
 		User user = UserRepository.updateMyProfile(data, loggedUser.getEmail());
 		if (user == null) {
@@ -125,6 +148,11 @@ public class UserService {
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response getByEmail(HashMap<String, String> data) {
+		loggedUser = (User) request.getSession().getAttribute("user");
+		
+		if (loggedUser == null) {
+			return Response.status(403).entity(HelperMethods.GetJsonValue("Unauthorized")).build();
+		}
 		User user = UserRepository.findByEmail(data.get("email"));
 		if (user == null)
 			return Response.status(400).entity("Error getting user").build();
@@ -136,9 +164,15 @@ public class UserService {
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response updateCategory(HashMap<String, String> data) {
+		
+		loggedUser = (User) request.getSession().getAttribute("user");
+		
+		if (loggedUser == null || loggedUser.getRole().equals(Role.user)) {
+			return Response.status(403).entity(HelperMethods.GetJsonValue("Unauthorized")).build();
+		}
 		if (!data.get("oldEmail").equals(data.get("email"))
 				&& !UserRepository.isUniqueEmail(data.get("email"))) {
-			return Response.status(200).entity("existError").build();
+			return Response.status(200).entity(HelperMethods.GetJsonValue("existError")).build();
 		}
 		boolean success = UserRepository.updateUser(data);
 		if (!success) {
@@ -152,9 +186,13 @@ public class UserService {
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response getbyOrganization() {
 		
-		User user = (User) request.getSession().getAttribute("user");
+		loggedUser = (User) request.getSession().getAttribute("user");
+		
+		if (loggedUser == null || loggedUser.getRole().equals(Role.superadmin)) {
+			return Response.status(403).entity(HelperMethods.GetJsonValue("Unauthorized")).build();
+		}
 
-		return Response.status(200).entity(UserRepository.getbyOrganization(user.getOrganization())).build();
+		return Response.status(200).entity(UserRepository.getbyOrganization(loggedUser.getOrganization())).build();
 
 	}
